Movida la actualización del UIState al Presenter
